refactor: move `google.mts` to the source files root
