+++ src/index.mts
@@ -2,4 +2,4 @@ export {
   googleGetWebsiteRank,
   googleListWebsites,
   GoogleWebsiteRank,
-} from "./utils/google.mjs";
+} from "./google.mjs";
