+++ src/utils/index.mts
@@ -1,3 +1,2 @@
 export { Arguments, ArgumentsParser } from "./arguments.mjs";
 export { formatKeywordRank } from "./format.mjs";
-export { googleGetWebsiteRank, GoogleWebsiteRank } from "./google.mjs";
